requirements.txt
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
folium==0.11.0
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
python-dotenv==0.14.0
ratelim==0.1.6
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
